# more faster api in wda for `typeKey` and `swipe`