--- a/.github/workflows/docker-image.yml
+++ b/.github/workflows/docker-image.yml
@@ -14,3 +14,5 @@ jobs:
     - uses: actions/checkout@v3
     - name: Build the Docker image
       run: docker build . --file Dockerfile --tag richard1ybb/stf:$(date +%s)
+    - name: Push the Docker image
+      run: docker push richard1ybb/stf:$(date +%s)