--- a/.github/workflows/docker-image.yml
+++ b/.github/workflows/docker-image.yml
@@ -17,4 +17,6 @@ jobs:
         tag: ${{ github.sha }}
       run: docker build . --file Dockerfile --tag richard1ybb/stf:$tag
     - name: Push the Docker image
+      env:
+        tag: ${{ github.sha }}
       run: docker push richard1ybb/stf:$tag