--- a/.github/workflows/docker-image.yml
+++ b/.github/workflows/docker-image.yml
@@ -15,7 +15,9 @@ jobs:
     - name: Build the Docker image
       env:
         tag: ${{ github.sha }}
-      run: docker build . --file Dockerfile --tag richard1ybb/stf:$tag
+      run: 
+        docker login --username=${{ secrets.DOCKER_USERNAME }} --password=${{ secrets.DOCKER_PASSWORD }}
+        docker build . --file Dockerfile --tag richard1ybb/stf:$tag
     - name: Push the Docker image
       env:
         tag: ${{ github.sha }}