--- a/.github/workflows/docker-image.yml
+++ b/.github/workflows/docker-image.yml
@@ -12,10 +12,16 @@ jobs:
 
     steps:
     - uses: actions/checkout@v3
+    - name: Login the Docker image
+      run:
+        docker login --username=${{ secrets.DOCKER_USERNAME }} --password=${{ secrets.DOCKER_PASSWORD }}
     - name: Build the Docker image
       env:
         tag: ${{ github.sha }}
-      run: |
-        docker login --username=${{ secrets.DOCKER_USERNAME }} --password=${{ secrets.DOCKER_PASSWORD }}
-        docker push richard1ybb/stf:$tag
+      run:
         docker build . --file Dockerfile --tag richard1ybb/stf:$tag
+    - name: Push the Docker image
+      env:
+        tag: ${{ github.sha }}
+      run:
+        docker push richard1ybb/stf:$tag