--- a/lib/units/ios-device/plugins/wda/WdaClient.js
+++ b/lib/units/ios-device/plugins/wda/WdaClient.js
@@ -134,7 +134,7 @@ module.exports = syrup.serial()
         this.sessionId = null
         return this.handleRequest({
           method: 'DELETE',
-          uri: `${this.baseUrl}/session/${currentSessionId}` 
+          uri: `${this.baseUrl}/session/${currentSessionId}`
         })
       },
       typeKey: function(params) {
@@ -208,13 +208,29 @@ module.exports = syrup.serial()
       setScaleByRotation: function(params, deviceSize) {
         return iosutil.swipe(this.orientation, params, deviceSize)
       },
+      getActions: function(params) {
+        return {
+          actions: [
+            {
+              type: 'pointer',
+              id: 'finger1',
+              parameters: {pointerType: 'touch'},
+              actions: [
+                {type: 'pointerMove', duration: 0, x: params.fromX, y: params.fromY},
+                {type: 'pointerMove', duration: params.duration, x: params.toX, y: params.toY},
+                {type: 'pointerUp'}
+              ]
+            }
+          ],
+        }
+      },
       swipe: function(params) {
         const scale = this.setScaleByRotation(params, this.deviceSize)
         this.isMove = true
         return this.handleRequest({
           method: 'POST',
-          uri: `${this.baseUrl}/session/${this.sessionId}/wda/dragfromtoforduration`,
-          body: scale,
+          uri: `${this.baseUrl}/session/${this.sessionId}/actions`,
+          body: this.getActions(scale),
           json: true
         })
       },