--- a/lib/units/ios-device/plugins/wda/WdaClient.js
+++ b/lib/units/ios-device/plugins/wda/WdaClient.js
@@ -217,7 +217,7 @@ module.exports = syrup.serial()
               parameters: {pointerType: 'touch'},
               actions: [
                 {type: 'pointerMove', duration: 0, x: params.fromX, y: params.fromY},
-                {type: 'pointerMove', duration: params.duration, x: params.toX, y: params.toY},
+                {type: 'pointerMove', duration: params.duration*1000, x: params.toX, y: params.toY},
                 {type: 'pointerUp'}
               ]
             }