`http://localhost:8100/session/27168052-6179-4684-89F2-2CB0160E9675/actions` use 636ms
`http://localhost:8100/session/27168052-6179-4684-89F2-2CB0160E9675/wda/dragfromtoforduration` use 1024ms

```
time curl --location --request POST 'http://localhost:8100/session/27168052-6179-4684-89F2-2CB0160E9675/actions' \
--header 'Content-Type: application/json' \
--data-raw '{
   "actions": [
     {
       "type": "pointer",
       "id": "finger1",
       "parameters": {"pointerType": "touch"},
       "actions": [
         {"type": "pointerMove", "duration": 0, "x": 160, "y": 200},
         {"type": "pointerMove", "duration": 100, "x": 360, "y": 200},
         {"type": "pointerUp"}
       ]
     }
   ]
 }'
{
  "value" : null,
  "sessionId" : "27168052-6179-4684-89F2-2CB0160E9675"
}curl --location --request POST  --header 'Content-Type: application/json'    0.01s user 0.01s system 2% cpu 0.636 total
```

```
time curl --location --request POST 'http://localhost:8100/session/27168052-6179-4684-89F2-2CB0160E9675/wda/dragfromtoforduration' \
--header 'Content-Type: application/json' \
--data-raw '{
    "duration": 0.1,
    "fromX": 160,
    "fromY": 200,
    "toX": 360,
    "toY": 200
}'
{
  "value" : null,
  "sessionId" : "27168052-6179-4684-89F2-2CB0160E9675"
}curl --location --request POST  --header 'Content-Type: application/json'    0.01s user 0.01s system 1% cpu 1.024 total
```