@vdelendik 2022-12-28
@danymill, let's replace oldschool `/wda/dragfromtoforduration` by `/actions`...